+++ EssentialFeed/EssentialFeedTests/Feed Cache/CacheFeedUseCaseTests.swift
@@ -148,42 +148,5 @@ class CacheFeedUseCaseTests: XCTestCase {
         URL(string: "http://any-url.com")!
     }
     
-    private func anyNSError() -> NSError { NSError(domain: "any error", code: 0) }
-    
-    private class FeedStoreSpy: FeedStore {
-        private var deletionCompletions = [DeletionCompletion]()
-        private var insertionCompletions = [InsertionCompletion]()
-        private(set) var receivedMessages = [ReceivedMessage]()
-        
-        enum ReceivedMessage: Equatable {
-            case deleteCachedFeed
-            case insert([LocalFeedImage], Date)
-        }
-        
-        func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-            deletionCompletions.append(completion)
-            receivedMessages.append(.deleteCachedFeed)
-        }
-        
-        func completeDeletion(with error: Error, at index: Int = 0) {
-            deletionCompletions[index](error)
-        }
-        
-        func completeDeletionSuccesfully(at index: Int = 0) {
-            deletionCompletions[index](nil)
-        }
-        
-        func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-            insertionCompletions.append(completion)
-            receivedMessages.append(.insert(feed, timestamp))
-        }
-        
-        func completeInsertion(with error: Error, at index: Int = 0) {
-            insertionCompletions[index](error)
-        }
-        
-        func completeInsertionSuccesfully(at index: Int = 0) {
-            insertionCompletions[index](nil)
-        }
-    }
+    private func anyNSError() -> NSError { NSError(domain: "any error", code: 0) }    
 }
